[feat] more secure secret storage
we want to avoid leaking secrets, so we don't include them in the source
files directly. even though we don't store the secret in the source at
time of writing, since the secret is eventually available at build time
in it's complete form (via ldflags injection), it can get pulled out of
the binary with the `strings` unix utility pretty easily by anyone who
knows the shape of the secret (e.g. hex string 32 bytes long).

instead of storing as a string, we do some code generation to make a go
source file that represents the secret and ID as a slice of 32 runes
that gets cast to a string:

     var clientID = string([]rune{'a', 'b', 'c', ...})

this prevents the secret from leaking out as a fully formed string that
can be pulled out by the `strings` utility