[wip] fuck it, just use urfave/cli
it's like 95% of what I want, it's tested, it has a similar API to what
I was thinking about and at this point I'd rather work on the "business
logic" of this side project than the plumbing. for the additional things
I want, I can probably use that library as a starting point and just
make some mods.